Struct mischief::Frame

pub struct Frame<T: Pointee + ?Sized, A: Allocator = Global>where
    T::Metadata: Metadata<T>,{ /* private fields */ }

A box that may be uninitialized.

Frames have the same memory layouts as Boxes.

Implementations

impl<T: Pointee + ?Sized, A: Allocator> Frame<T, A>where
T::Metadata: Metadata<T>,

pub fn allocator(f: &Self) -> &A

Returns a reference to the underlying allocator.

Note: this is an associated function, which means that you have to call it as Frame::allocator(&f) instead of f.allocator(). This is so that there is no conflict with a method on the inner type.

pub fn as_ptr(&self) -> *const T

Returns a pointer to the underlying memory.

pub fn as_mut_ptr(&mut self) -> *mut T

Returns a mutable pointer to the underlying memory.

pub unsafe fn assume_init(self) -> Box<T, A>

Converts to Box<T, A>.

This does not move the value pointed to by the frame.

Safety

The contents of the frame must be initialized. Calling this when the content is not yet initialized causes immediate undefined behavior.

Additionally, most types have additional invariants beyond merely being considered initialized at the type level. For example, a 1-initialized Vec<T> is considered initialized (under the current implementation; this does not constitute a stable guarantee) because the only requirement the compiler knows about it is that the data pointer must be non-null. Creating such a Vec<T> does not cause immediate undefined behavior, but will cause undefined behavior with most safe operations (including dropping it).

pub unsafe fn from_raw_in(raw: *mut T, alloc: A) -> Self

Constructs a frame from a raw pointer in the given allocator.

After calling this function, the raw pointer is owned by the resulting Frame. Specifically, the Frame destructor will free the allocated memory. For this to be safe, the memory must have been allocated in accordance with the memory layout used by Frame.

Safety

raw must be non-null and allocated by alloc according to the memory layout used by Frame.

pub fn into_raw_with_allocator(f: Self) -> (*mut T, A)

Consumes the Frame, returning a raw pointer and the allocator.

The pointer will be properly aligned and non-null.

After calling this function, the caller is responsible for the memory previously managed by the Frame. In particular, the caller should properly release the memory, taking into account the memory layout used by Frame. The easiest way to do this is to convert the raw pointer back into a Frame with the Frame::from_raw_in function, allowing the Frame destructor to perform the cleanup.

Note: this is an associated function, which means that you have to call it as Frame::into_raw_with_allocator(f) instead of f.into_raw_with_allocator(). This is so that there is no conflict with a method on the inner type.

pub unsafe fn new_unsized_in(metadata: T::Metadata, alloc: A) -> Self

Allocates memory for an unsized type with the given metadata in the given allocator.

This doesn’t actually allocate if the metadata provides a layout with zero size.

Safety

metadata must be valid for a pointer to T.

pub fn slot(&mut self) -> Slot<'_, T>

Returns a Slot of the internal contents.

impl<T, A: Allocator> Frame<T, A>

pub fn into_framed_slice(framed: Self) -> Frame<[T], A>

Converts a Frame<T> into a Frame<[T]>.

This conversion does not allocate on the heap and happens in place.

pub fn new_in(alloc: A) -> Self

Allocates memory in the given allocator.

This doesn’t actually allocate if T is zero-sized.

pub fn init(self, value: T) -> Box<T, A>

Sets the value of the underlying memory and converts the frame to a box.

This overwrites any previous value without dropping it, so be careful not to use this after initializing the frame unless you want to skip running the destructor.

impl<T: Pointee + ?Sized> Frame<T, Global>where
T::Metadata: Metadata<T>,

pub unsafe fn from_raw(raw: *mut T) -> Self

Constructs a frame from a raw pointer.

After calling this function, the raw pointer is owned by the resulting Frame. Specifically, the Frame destructor will free the allocated memory. For this to be safe, the memory must have been allocated in accordance with the memory layout used by Frame.

Safety

This function is unsafe because improper use may lead to memory problems. For example, a double-free may occur if the function is called twice on the same raw pointer.

The safety conditions are described in the memory layout section.

pub fn into_raw(f: Self) -> *mut T

Consumes the Frame, returning a raw pointer.

The pointer will be properly aligned and non-null.

After calling this function, the caller is responsible for the memory previously managed by the Frame. In particular, the caller should properly release the memory, taking into account the memory layout used by Frame. The easiest way to do this is to convert the raw pointer back into a Frame with the Frame::from_raw function, allowing the Frame destructor to perform the cleanup.

Note: this is an associated function, which means that you have to call it as Frame::into_raw(f) instead of f.into_raw(). This is so that there is no conflict with a method on the inner type.

pub unsafe fn new_unsized(metadata: T::Metadata) -> Self

Allocates memory for an object with the given metadata on the heap.

This doesn’t actually allocate if the metadata provides a layout with zero size.

Safety

metadata must be valid for a pointer to T.

impl<T> Frame<T, Global>

pub fn new() -> Self

Allocates memory on the heap.

This doesn’t actually allocate if T is zero-sized.