Struct mischief::In

pub struct In<P, R: Region> { /* private fields */ }

A pointer which has its pointee in a specific memory region.

Implementations

impl<P: Pointer, R: Region> In<P, R>where
P::Target: Pointee,

pub fn new(ptr: P) -> Selfwhere
P: Within<R>,

Creates a new In from a pointer.

pub unsafe fn new_unchecked(ptr: P) -> Self

Creates a new In from a pointer.

Safety

The pointee of ptr must be contained in R.

pub fn map<F, Q>(self, f: F) -> In<Q, R>where
F: FnOnce(P) -> Q,
Q: Pointer + Within<R>,
Q::Target: Pointee,

Maps this In to another pointer in its region.

pub unsafe fn map_unchecked<F, Q>(self, f: F) -> In<Q, R>where
F: FnOnce(P) -> Q,
Q: Pointer,
Q::Target: Pointee,

Maps this In to another pointer in its region.

Safety

The pointer returned by f must be completely contained in R.

pub fn as_raw(&self) -> In<*mut P::Target, R>

Gets a raw In from this pointer.

impl<P, R: Region> In<P, R>

pub fn into_inner(this: Self) -> P

Unwraps an In, returning the underlying pointer.

pub fn ptr(&self) -> &P

Returns a reference to the pointer of this In.

pub unsafe fn ptr_mut(&mut self) -> &mut P

Returns a mutable reference to the pointer of this In.

Safety

The pointer must not be mutated to point outside of R.

impl<T: Pointee + ?Sized, A: RegionalAllocator> In<Frame<T, A>, A::Region>where
<T as Pointee>::Metadata: Metadata<T>,

pub fn slot(&mut self) -> In<Slot<'_, T>, A::Region>

Returns a Slot of the internal contents.

impl<'a, T: Pointee + ?Sized, R: Region> In<Slot<'a, T>, R>

pub fn as_mut<'b>(&'b mut self) -> In<Slot<'b, T>, R>where
'a: 'b,

Gets a mutable borrow from this slot.