Struct mischief::Slot

pub struct Slot<'a, T: ?Sized> { /* private fields */ }

A memory location that may or may not have a value initialized in it.

Implementations

impl<'a, T: ?Sized> Slot<'a, T>

pub fn pointer_layout(&self) -> Layoutwhere
T: Pointee,
T::Metadata: Metadata<T>,

Returns the layout of the pointer held in the slot.

pub unsafe fn new_unchecked(ptr: *mut T) -> Self

Creates a new slot from an exclusive pointer.

Safety

ptr must be non-null, properly aligned, and valid for reads and writes.

ptr must not alias any other accessible references for 'a.

pub unsafe fn assume_init_ref(&self) -> &'a T

Gets a shared reference to the value in the slot.

Safety

The contents of the slot must be initialized. See assume_init for details.

pub unsafe fn assume_init_mut(&mut self) -> &'a mut T

Gets a mutable reference to the value in the slot.

Safety

The contents of the slot must be initialized. See assume_init for details.

pub unsafe fn assume_init_drop(&mut self)

Drops the value in the slot.

Safety

The contents of the slot must be initialized. See assume_init for details.

pub fn as_ptr(&self) -> *mut T

Gets a pointer to the underlying memory.

pub fn as_mut<'b>(&'b mut self) -> Slot<'b, T>where
'a: 'b,

Gets a mutable borrow from this slot.

pub fn zero(&mut self)where
T: Pointee,
T::Metadata: Metadata<T>,

Writes zeroes to every byte of the backing memory.

pub fn as_bytes(self) -> Slot<'a, [u8]>where
T: Pointee,
T::Metadata: Metadata<T>,

Returns a slot of the underlying bytes.

pub fn cast_checked<U>(self) -> Option<Slot<'a, U>>where
T: Pointee,
T::Metadata: Metadata<T>,

Attempts to cast the type of the Slot from T to U.

Returns None if U requires a higher alignment than the pointer in self or U is larger than T.

pub unsafe fn cast<U>(self) -> Slot<'a, U>

Casts the type of the Slot from T to U.

Safety

The slot must point to memory suitable for holding a U. In addition to being the right size and aligment, the caller must also guarantee that the slot pointer will not alias any other accessible references after being cast.

impl<'a, T> Slot<'a, T>

pub fn new(value: &'a mut MaybeUninit<T>) -> Self

Returns a new Slot backed by the given MaybeUninit.

pub fn write(&mut self, value: T) -> &mut T

Sets the value of the Slot.

This overwrites any previous value without dropping it, so be careful not to use this after initializing the slot unless you want to skip running the destructor.

Because this method does not convert the Slot to a value, the written value will not be dropped when the returned reference or underlying Slot leave scope.

pub fn as_maybe_uninit(&self) -> &MaybeUninit<T>

Returns a reference to the underlying memory as a MaybeUninit.

pub fn as_maybe_uninit_mut(&mut self) -> &mut MaybeUninit<T>

Returns a reference to the underlying memory as a mutable MaybeUninit.

impl<'a, T, const N: usize> Slot<'a, [T; N]>

pub fn len(&self) -> usize

Returns the length of the pointed-to array.

pub fn is_empty(&self) -> bool

Returns whether the pointed-to array is empty.

pub fn get(self, index: usize) -> Slot<'a, T>

Gets an element of the slot array.

pub unsafe fn get_unchecked(self, index: usize) -> Slot<'a, T>

Gets an element of the slot array without doing bounds checking.

For a safe alternative, see get.

Safety

Calling this method with an out-of-bounds index is undefined behavior even if the resulting reference is not used.

pub fn unsize(self) -> Slot<'a, [T]>

Converts an slot array to an slot slice of the same length.

impl<'a, T> Slot<'a, [T]>

pub fn len(&self) -> usize

Returns the length of the pointed-to slice.

pub fn is_empty(&self) -> bool

Returns whether the pointed-to slice is empty.

pub fn get(self, index: usize) -> Slot<'a, T>

Gets an element of the slot slice.

pub unsafe fn get_unchecked(self, index: usize) -> Slot<'a, T>

Gets an element of the slot slice without doing bounds checking.

For a safe alternative, see get.

Safety

Calling this method with an out-of-bounds index is undefined behavior even if the resulting reference is not used.