Trait mischief::Metadata

pub unsafe trait Metadata<T: Pointee<Metadata = Self> + ?Sized>: Copy + Send + Sync + Ord + Hash + Unpin {
    unsafe fn pointee_layout(self) -> Layout;
}

Pointer metadata that can determine the memory layout of its pointee.

Safety

pointee_layout must return the correct layout of a T pointee with this metadata.

Required Methods

unsafe fn pointee_layout(self) -> Layout

Returns the layout of a T pointee with the this metadata.

Safety

self must be valid metadata for T.