Trait mischief::Pointer

pub unsafe trait Pointer {
    type Target: ?Sized;

    fn target(&self) -> *mut Self::Target;
}

A type that points to a single location in memory.

Safety

If the type implementing Pointer also implements Deref or DerefMut, the pointer returned by Target must be equal to the references returned from deref and deref_mut.

Required Associated Types

type Target: ?Sized

The target value of this type.

Required Methods

fn target(&self) -> *mut Self::Target

Returns a pointer to this type’s target.