Trait mischief::RegionalAllocator

pub unsafe trait RegionalAllocator: Allocator {
    type Region: Region;
}

An Allocator that allocates inside a single contiguous memory region.

Safety

The pointers returned from a RegionalAllocator’s Allocator implementation must always be contained in its associated Region.

Required Associated Types

type Region: Region

The region type for this allocator.