Trait mischief::Static

pub trait Static {
    type Unique: Unique;
    type Target: 'static;

    unsafe fn slot() -> Slot<'static, Self::Target>;
}

A type that can lease access to a type without any context.

Required Associated Types

type Unique: Unique

The unique type that can be used to lease this static memory location.

type Target: 'static

The type that a reference can be created to.

Required Methods

unsafe fn slot() -> Slot<'static, Self::Target>

Returns a mutable reference to a slot of the target type.

Safety

The caller must hold a mutable borrow of the Unique.